move testconfig tests to use appropriate method
